pyproject.toml
[project]
name = "pmsmanagebot"
version = "0.4.1"
description = "Telegram bot for managing Plex/Emby media servers with WebApp support"
authors = [
    { name = "WithdewHua", email = "i@10101.io" }
]
dependencies = [
    "plexapi==4.13.0",
    "pydantic-settings>=2.7.0",
    "python-telegram-bot==20.0a2",
    "redis>=5.2.1",
    "requests==2.28.0",
    "urllib3==1.26.9",
    "fastapi==0.103.0",
    "uvicorn==0.23.2",
    "gunicorn>=21.0.0",
    "python-multipart==0.0.6",
    "itsdangerous>=2.2.0",
    "pydantic[email]>=2.10.4",
    "aiohttp>=3.11.18",
    "filelock>=3.18.0",
    "sqlalchemy>=1.4.0,<2.0.0",
    "slowapi>=0.1.9",
]
readme = "README.md"
requires-python = ">= 3.11"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/app"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
